🚚  separate `Page` from `Layout`; canonical instead of url
