+++ sites/jeromefitzgerald.com/src/components/Layout/index.ts
@@ -1,5 +1,1 @@
-import Layout from './Layout'
-import Page from './Page'
-
-export { Layout, Page }
-export default Layout
+export { Layout } from './Layout'
