+++ sites/jeromefitzgerald.com/src/components/Page/index.ts
@@ -1,0 +1,1 @@
+export { Page } from './Page'
